Extract tests to useBusReducer and useBusState
