+++ src/react.useBusReducer.test.tsx
@@ -1,37 +1,14 @@
 import { act, renderHook } from "@testing-library/react-hooks";
-import { EventEmitter2 } from "eventemitter2";
+
 import React from "react";
-import { create } from "react-test-renderer";
-import { createEventDefinition, EventBus } from "./EventBus";
-import {
-  BusProvider,
-  useBus,
-  useBusReducer,
-  useBusState,
-  stateSubscriber,
-  reducerSubscriber
-} from "./react";
+
+import { BusProvider, useBusReducer, reducerSubscriber } from "./react";
 import { SubscribeFn, BusEvent } from "./types";
 import { _defaultSubscriber } from "./useBusReducer";
-
-const bus = new EventBus();
-
-function mockEventBus() {
-  const _unsubscribe = jest.fn();
-  const subscribe = jest.fn(() => _unsubscribe);
-  const publish = jest.fn();
-  const emitter = new EventEmitter2();
-  return { subscribe, emitter, publish, _unsubscribe };
-}
-
-const wrapper = ({ children }: { children?: React.ReactNode }) => (
-  <BusProvider value={bus}>{children}</BusProvider>
-);
-
-it("should provide a bus", () => {
-  const { result } = renderHook(() => useBus(), { wrapper });
-  expect(result.current).toBe(bus);
-});
+import { mockEventBus, bus, wrapper } from "./testhelpers";
+import { createEventDefinition, EventBus } from "./EventBus";
+import { create } from "react-test-renderer";
+import { useBus } from "./BusContext";
 
 it("should not subscribe without unsubscribing (useBusReducer)", () => {
   const mockBus = mockEventBus();
@@ -60,86 +37,6 @@ it("should not subscribe without unsubscribing (useBusReducer)", () => {
 
   expect(mockBus.subscribe.mock.calls.length).toBe(2);
   expect(mockBus._unsubscribe.mock.calls.length).toBe(2);
-});
-
-it("should not subscribe without unsubscribing (useBusState)", () => {
-  const mockBus = mockEventBus();
-  const incrementEvent = createEventDefinition<number>()("increment");
-
-  // run once to subscribe to bus
-  const hook = renderHook(() => useBusState(0, incrementEvent), {
-    wrapper: ({ children }: { children?: React.ReactNode }) => (
-      <BusProvider value={mockBus}>{children}</BusProvider>
-    )
-  });
-
-  hook.unmount();
-
-  expect(mockBus.subscribe.mock.calls.length).toBe(1);
-  expect(mockBus._unsubscribe.mock.calls.length).toBe(1);
-});
-
-it("should update state (options configuration)", () => {
-  const incrementEvent = createEventDefinition<number>()("counter.increment");
-
-  const { result } = renderHook(
-    () =>
-      useBusState.configure({
-        subscriber: (dispatch, bus) => {
-          return bus.subscribe("counter.**", v => dispatch(v.payload));
-        }
-      })(0),
-    {
-      wrapper
-    }
-  );
-
-  expect(result.current).toBe(0);
-
-  act(() => {
-    bus.publish(incrementEvent(1));
-  });
-
-  expect(result.current).toBe(1);
-});
-
-it("should update state by subscribing to multiple events", () => {
-  const positiveNumberEvent = createEventDefinition<number>()("positive");
-  const negativeNumberEvent = createEventDefinition<number>()("negative");
-
-  const { result } = renderHook(
-    () =>
-      useBusState.configure({
-        subscriber: stateSubscriber("positive", "negative")
-      })(0),
-    {
-      wrapper
-    }
-  );
-
-  expect(result.current).toBe(0);
-
-  act(() => bus.publish(positiveNumberEvent(15)));
-  expect(result.current).toBe(15);
-
-  act(() => bus.publish(negativeNumberEvent(-5)));
-  expect(result.current).toBe(-5);
-});
-
-it("should update state", () => {
-  const incrementEvent = createEventDefinition<number>()("increment");
-
-  const { result } = renderHook(() => useBusState(0, incrementEvent), {
-    wrapper
-  });
-
-  expect(result.current).toBe(0);
-
-  act(() => {
-    bus.publish(incrementEvent(1));
-  });
-
-  expect(result.current).toBe(1);
 });
 
 it("should not reduce for not subscribed event", () => {
